pyproject.toml
[project]
name = "poll-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.30.0",
    "discord-py>=2.6.2",
    "python-dotenv>=1.1.1",
]
